fix: add protection for non owner

+++ messages/delete.php
@@ -9,6 +9,12 @@
     require '../utils/functions.php';
 
     $dId = $_GET['id'];
+    $message = getByQuery("SELECT * FROM messages WHERE id = $dId")[0];
+
+    if ($message["user_id"] != $_SESSION["uid"]) {
+        header('Location: index.php');
+        exit;
+    }
 
     if (deleteMsg($dId) > 0) {
         echo "
